+++ changelog.md
@@ -75,6 +75,7 @@
 - Import of files during WooCommerce CSV supports now also usage of credentials (you could import files e.g. from FTP)
 - Using new transient object in backend for hints and errors
 - Cleanup the return value for external files via get_attached_file()
+- File protocol uses now WP_Filesystem for each file interaction
 
 ### Fixed